requirements.txt.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
gunicorn==20.1.0